// src/models/Requisiciones/probatory.ts
export type periodo_prueba = {
  id: string;
  user_email: string;
  user_name: string;
  user_document: string;
  candidate_adaptation: string;
  candidate_performance: string;
  was_feedback: boolean;
  had_candidate_asked: boolean;
  had_candidate_improve_area: boolean;
  was_explication_about_role: boolean;
  had_candidate_achive_probatory_goals: string;
  hora_inicio: Date
}